+++ src/Translator/src/Catalogue/CatalogueManager.php
@@ -1,12 +1,5 @@
 <?php
 
-/**
- * Spiral Framework.
- *
- * @license   MIT
- * @author    Anton Titov (Wolfy-J)
- */
-
 declare(strict_types=1);
 
 namespace Spiral\Translator\Catalogue;
@@ -20,30 +13,21 @@ use Spiral\Translator\CatalogueManagerInterface;
  */
 final class CatalogueManager implements CatalogueManagerInterface
 {
-    /** @var LoaderInterface */
-    private $loader;
+    private array $locales = [];
 
-    /**
-     * @internal
-     * @var CacheInterface
-     */
-    private $cache;
-
-    /** @var array */
-    private $locales = [];
+    /** @internal */
+    private readonly CacheInterface $cache;
 
     /** @var Catalogue[] */
-    private $catalogues = [];
+    private array $catalogues = [];
 
-    public function __construct(LoaderInterface $loader, CacheInterface $cache = null)
-    {
-        $this->loader = $loader;
+    public function __construct(
+        private readonly LoaderInterface $loader,
+        CacheInterface $cache = null
+    ) {
         $this->cache = $cache ?? new NullCache();
     }
 
-    /**
-     * @inheritdoc
-     */
     public function getLocales(): array
     {
         if ($this->locales !== []) {
@@ -59,9 +43,6 @@ final class CatalogueManager implements CatalogueManagerInterface
         return $this->locales;
     }
 
-    /**
-     * @inheritdoc
-     */
     public function load(string $locale): CatalogueInterface
     {
         if (isset($this->catalogues[$locale])) {
@@ -78,25 +59,16 @@ final class CatalogueManager implements CatalogueManagerInterface
         return $this->catalogues[$locale];
     }
 
-    /**
-     * @inheritdoc
-     */
     public function save(string $locale): void
     {
         $this->cache->saveLocale($locale, $this->get($locale)->getData());
     }
 
-    /**
-     * @inheritdoc
-     */
     public function has(string $locale): bool
     {
-        return isset($this->catalogues[$locale]) || in_array($locale, $this->getLocales());
+        return isset($this->catalogues[$locale]) || \in_array($locale, $this->getLocales());
     }
 
-    /**
-     * @inheritdoc
-     */
     public function get(string $locale): CatalogueInterface
     {
         return $this->load($locale);