+++ src/Translator/src/Indexer.php
@@ -1,12 +1,5 @@
 <?php
 
-/**
- * Spiral Framework.
- *
- * @license   MIT
- * @author    Anton Titov (Wolfy-J)
- */
-
 declare(strict_types=1);
 
 namespace Spiral\Translator;
@@ -32,20 +25,13 @@ final class Indexer
 {
     use LoggerTrait;
 
-    /** @var TranslatorConfig */
-    private $config;
-
     /**
-     * Catalogue to aggregate messages into.
-     *
-     * @var CatalogueInterface
+     * @param CatalogueInterface $catalogue Catalogue to aggregate messages into.
      */
-    private $catalogue;
-
-    public function __construct(TranslatorConfig $config, CatalogueInterface $catalogue)
-    {
-        $this->config = $config;
-        $this->catalogue = $catalogue;
+    public function __construct(
+        private readonly TranslatorConfig $config,
+        private readonly CatalogueInterface $catalogue
+    ) {
     }
 
     /**
@@ -61,8 +47,8 @@ final class Indexer
         $this->catalogue->set($domain, $string, $string);
 
         $this->getLogger()->debug(
-            sprintf('[%s]: `%s`', $domain, $string),
-            compact('domain', 'string')
+            \sprintf('[%s]: `%s`', $domain, $string),
+            ['domain' => $domain, 'string' => $string]
         );
     }
 
@@ -121,22 +107,20 @@ final class Indexer
 
     /**
      * Fetch default string values from class and merge it with parent strings if requested.
-     *
-     * @return array
      */
-    private function fetchMessages(\ReflectionClass $reflection, bool $inherit = false)
+    private function fetchMessages(\ReflectionClass $reflection, bool $inherit = false): array
     {
         $target = $reflection->getDefaultProperties() + $reflection->getConstants();
 
         foreach ($reflection->getProperties() as $property) {
-            if (is_string($property->getDocComment()) && strpos($property->getDocComment(), '@do-not-index')) {
+            if (\is_string($property->getDocComment()) && \strpos($property->getDocComment(), '@do-not-index')) {
                 unset($target[$property->getName()]);
             }
         }
 
         $strings = [];
-        array_walk_recursive($target, function ($value) use (&$strings): void {
-            if (is_string($value) && Translator::isMessage($value)) {
+        \array_walk_recursive($target, function ($value) use (&$strings): void {
+            if (\is_string($value) && Translator::isMessage($value)) {
                 $strings[] = $this->prepareMessage($value);
             }
         });
@@ -144,7 +128,7 @@ final class Indexer
         if ($inherit && $reflection->getParentClass()) {
             //Joining strings data with parent class values (inheritance ON) - resolved into same
             //domain on export
-            $strings = array_merge(
+            $strings = \array_merge(
                 $strings,
                 $this->fetchMessages($reflection->getParentClass(), true)
             );
@@ -167,19 +151,11 @@ final class Indexer
         }
 
         //`l` and `p`, `say` functions
-        $argument = null;
-        switch (strtolower($invocation->getName())) {
-            case 'say':
-            case 'l':
-                if ($invocation->countArguments() >= 3) {
-                    $argument = $invocation->getArgument(2);
-                }
-                break;
-            case 'p':
-                if ($invocation->countArguments() >= 4) {
-                    $argument = $invocation->getArgument(3);
-                }
-        }
+        $argument = match (\strtolower($invocation->getName())) {
+            'say', 'l' => $invocation->countArguments() >= 3 ? $invocation->getArgument(2) : null,
+            'p' => $invocation->countArguments() >= 4 ? $invocation->getArgument(3) : null,
+            default => null
+        };
 
         if (!empty($argument) && $argument->getType() === ReflectionArgument::STRING) {
             //Domain specified in arguments
@@ -195,7 +171,7 @@ final class Indexer
     private function prepareMessage(string $string): string
     {
         if (Translator::isMessage($string)) {
-            $string = substr($string, 2, -2);
+            $string = \substr($string, 2, -2);
         }
 
         return $string;