+++ src/Translator/src/Matcher.php
@@ -1,12 +1,5 @@
 <?php
 
-/**
- * Spiral Framework.
- *
- * @license   MIT
- * @author    Anton Titov (Wolfy-J)
- */
-
 declare(strict_types=1);
 
 namespace Spiral\Translator;
@@ -20,7 +13,7 @@ final class Matcher
 {
     public function isPattern(string $string): bool
     {
-        return strpos($string, '*') !== false || strpos($string, '|') !== false;
+        return \strpos($string, '*') !== false || \strpos($string, '|') !== false;
     }
 
     /**
@@ -28,19 +21,16 @@ final class Matcher
      */
     public function matches(string $string, string $pattern): bool
     {
-        if ($string === $pattern) {
-            return true;
-        }
-        if (!$this->isPattern($pattern)) {
-            return false;
-        }
-
-        return (bool)preg_match($this->getRegex($pattern), $string);
+        return match (true) {
+            $string === $pattern => true,
+            !$this->isPattern($pattern) => false,
+            default => (bool) \preg_match($this->getRegex($pattern), $string)
+        };
     }
 
     private function getRegex(string $pattern): string
     {
-        $regex = str_replace('*', '[a-z0-9_\-]+', addcslashes($pattern, '.-'));
+        $regex = \str_replace('*', '[a-z0-9_\-]+', \addcslashes($pattern, '.-'));
 
         return "#^{$regex}$#i";
     }